Disable Academic theme thumbnail
A week ago I opened  https://github.com/gcushen/hugo-academic/issues/1684 to request the removal of a slogan and device mock-ups from the theme's thumbnail. The author has not addressed the issue one week later. Therefore with this commit I am disabling `tn.png` for the Academic theme and instead use its `screenshot.png` (that does not contain the slogan and the mock-ups) for the Themes Site. 

Whenever the author responds this commit will be reverted. 

cc: @bep @digitalcraftsman